create new metadata object on mobs, when metadata definitions are included in yaml
referenced here: https://github.com/RanvierMUD/ranviermud/issues/390
Currently if you define any metadata in an npcs.yml file, then all copies of that npc share a single metadata object. Not ideal! Being able to set metadata in yaml is quite useful though, and it would be a shame to not be able to do that.